Refactor: Moved ChunkFormat from SliceView to SliceView/Volume
ChunkFormat is specific to VolumeChunks, which render chunk data directly to textures.
